+++ src/components/output/StageView.svelte
@@ -221,10 +221,10 @@
         untrack(() => resetAnimator());
     });
 
-    // Stop or When the evaluator is playing but the animator is stopped, create a new animator.
+    // When the evaluator is playing but the animator is stopped, create a new animator.
     $effect(() => {
         if (animator) {
-            if ($evaluation.playing) {
+            if ($evaluation?.playing) {
                 if (animator.isStopped()) untrack(() => resetAnimator());
             } else {
                 animator.stop();